+++ src/utils.js
@@ -1,6 +1,6 @@
 import { URI } from 'stac-js/src/utils.js';
 import removeMd from 'remove-markdown';
-import { Link, Asset } from 'stac-js';
+import { Link } from 'stac-js';
 import { hasText, isObject, size } from 'stac-js/src/utils.js';
 import { geojsonMediaType, imageMediaTypes } from 'stac-js/src/mediatypes.js';
 import { pagination } from "stac-js/src/relationtypes.js";
@@ -265,6 +265,16 @@ export default class Utils {
     }
   }
 
+  static getIcon(data) {
+    if (data?.isSTAC) {
+      const icons = data.getIcons();
+      if (icons.length > 0) {
+        return icons[0];
+      }
+    }
+    return null;
+  }
+
   static titleForHref(href, preferFileName = false) {
     const uri = URI(href);
     const auth = uri.authority();
@@ -376,7 +386,7 @@ export default class Utils {
 
   static assetFilename(asset, response = null) {
     // Get the preferred filename from the file:local_path property
-    if (asset instanceof Asset) {
+    if (asset?.isAsset) {
       const localPath = asset.getMetadata('file:local_path');
       if (typeof localPath === 'string') {
         return URI(localPath).filename();